pyproject.toml
[tool.poetry]
name = "soundx"
version = "0.1.0"
description = ""
authors = ["Julien Roullé <ju.roulle@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pandas = "^2.1.2"
streamlit = "^1.28.1"
boto3 = "^1.28.77"
matplotlib = "^3.8.1"
librosa = "^0.10.1"
python-dotenv = "^1.0.0"
tensorflow-hub = "^0.15.0"
seaborn = "^0.13.0"
datasets = "^2.15.0"
evaluate = "^0.4.1"
ipykernel = "^6.27.1"
ipywidgets = "^8.1.1"
tensorflow = "^2.15.0"
modal = "^0.63.98"
pydub = "^0.25.1"

[tool.poetry.group.dev.dependencies]
tensorflow-macos = "2.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
